# Make putFixed{32,64} more strict, speeding up encoding of floats.