Further improvements over #303.  The standard functions from
`Data.ByteString.Builder` aren't strict enough and miss some potential
unwrapping opportunities.

Benchmarks (on top of the `internal-vectors` branch):

float-packed(4003B)/encode: 7.346 μs => 3.652 μs
float-unpacked(5000B)/encode: 17.13 μs => 12.64 μs

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/google/proto-lens/304)
<!-- Reviewable:end -->